+++ compiler/testData/codegen/boxWithStdlib/ranges/expression/inexactSteppedDownTo.kt
@@ -12,23 +12,23 @@ fun box(): String {
         return "Wrong elements for 8 downTo 3 step 2: $list1"
     }
 
-    val list2 = ArrayList<Byte>()
+    val list2 = ArrayList<Int>()
     val range2 = 8.toByte() downTo 3.toByte() step 2
     for (i in range2) {
         list2.add(i)
         if (list2.size() > 23) break
     }
-    if (list2 != listOf<Byte>(8, 6, 4)) {
+    if (list2 != listOf<Int>(8, 6, 4)) {
         return "Wrong elements for 8.toByte() downTo 3.toByte() step 2: $list2"
     }
 
-    val list3 = ArrayList<Short>()
+    val list3 = ArrayList<Int>()
     val range3 = 8.toShort() downTo 3.toShort() step 2
     for (i in range3) {
         list3.add(i)
         if (list3.size() > 23) break
     }
-    if (list3 != listOf<Short>(8, 6, 4)) {
+    if (list3 != listOf<Int>(8, 6, 4)) {
         return "Wrong elements for 8.toShort() downTo 3.toShort() step 2: $list3"
     }
