+++ compiler/testData/codegen/boxWithStdlib/ranges/literal/oneElementRange.kt
@@ -11,21 +11,21 @@ fun box(): String {
         return "Wrong elements for 5..5: $list1"
     }
 
-    val list2 = ArrayList<Byte>()
+    val list2 = ArrayList<Int>()
     for (i in 5.toByte()..5.toByte()) {
         list2.add(i)
         if (list2.size() > 23) break
     }
-    if (list2 != listOf<Byte>(5.toByte())) {
+    if (list2 != listOf<Int>(5)) {
         return "Wrong elements for 5.toByte()..5.toByte(): $list2"
     }
 
-    val list3 = ArrayList<Short>()
+    val list3 = ArrayList<Int>()
     for (i in 5.toShort()..5.toShort()) {
         list3.add(i)
         if (list3.size() > 23) break
     }
-    if (list3 != listOf<Short>(5.toShort())) {
+    if (list3 != listOf<Int>(5)) {
         return "Wrong elements for 5.toShort()..5.toShort(): $list3"
     }
