+++ compiler/testData/codegen/boxWithStdlib/ranges/literal/reversedBackSequence.kt
@@ -11,21 +11,21 @@ fun box(): String {
         return "Wrong elements for (5 downTo 3).reversed(): $list1"
     }
 
-    val list2 = ArrayList<Byte>()
+    val list2 = ArrayList<Int>()
     for (i in (5.toByte() downTo 3.toByte()).reversed()) {
         list2.add(i)
         if (list2.size() > 23) break
     }
-    if (list2 != listOf<Byte>(3, 4, 5)) {
+    if (list2 != listOf<Int>(3, 4, 5)) {
         return "Wrong elements for (5.toByte() downTo 3.toByte()).reversed(): $list2"
     }
 
-    val list3 = ArrayList<Short>()
+    val list3 = ArrayList<Int>()
     for (i in (5.toShort() downTo 3.toShort()).reversed()) {
         list3.add(i)
         if (list3.size() > 23) break
     }
-    if (list3 != listOf<Short>(3, 4, 5)) {
+    if (list3 != listOf<Int>(3, 4, 5)) {
         return "Wrong elements for (5.toShort() downTo 3.toShort()).reversed(): $list3"
     }
