+++ compiler/testData/codegen/boxWithStdlib/ranges/literal/reversedEmptyBackSequence.kt
@@ -11,21 +11,21 @@ fun box(): String {
         return "Wrong elements for (3 downTo 5).reversed(): $list1"
     }
 
-    val list2 = ArrayList<Byte>()
+    val list2 = ArrayList<Int>()
     for (i in (3.toByte() downTo 5.toByte()).reversed()) {
         list2.add(i)
         if (list2.size() > 23) break
     }
-    if (list2 != listOf<Byte>()) {
+    if (list2 != listOf<Int>()) {
         return "Wrong elements for (3.toByte() downTo 5.toByte()).reversed(): $list2"
     }
 
-    val list3 = ArrayList<Short>()
+    val list3 = ArrayList<Int>()
     for (i in (3.toShort() downTo 5.toShort()).reversed()) {
         list3.add(i)
         if (list3.size() > 23) break
     }
-    if (list3 != listOf<Short>()) {
+    if (list3 != listOf<Int>()) {
         return "Wrong elements for (3.toShort() downTo 5.toShort()).reversed(): $list3"
     }
