+++ compiler/testData/codegen/boxWithStdlib/ranges/literal/simpleDownTo.kt
@@ -11,21 +11,21 @@ fun box(): String {
         return "Wrong elements for 9 downTo 3: $list1"
     }
 
-    val list2 = ArrayList<Byte>()
+    val list2 = ArrayList<Int>()
     for (i in 9.toByte() downTo 3.toByte()) {
         list2.add(i)
         if (list2.size() > 23) break
     }
-    if (list2 != listOf<Byte>(9, 8, 7, 6, 5, 4, 3)) {
+    if (list2 != listOf<Int>(9, 8, 7, 6, 5, 4, 3)) {
         return "Wrong elements for 9.toByte() downTo 3.toByte(): $list2"
     }
 
-    val list3 = ArrayList<Short>()
+    val list3 = ArrayList<Int>()
     for (i in 9.toShort() downTo 3.toShort()) {
         list3.add(i)
         if (list3.size() > 23) break
     }
-    if (list3 != listOf<Short>(9, 8, 7, 6, 5, 4, 3)) {
+    if (list3 != listOf<Int>(9, 8, 7, 6, 5, 4, 3)) {
         return "Wrong elements for 9.toShort() downTo 3.toShort(): $list3"
     }
